solve 2.5
